fix circular imports for typehints
