+++ src/app/pieces.py
@@ -1,14 +1,13 @@
-import re
 from math import inf as infinity
-from string import ascii_lowercase
-from typing import Union, Tuple, List, Iterator
+from typing import TYPE_CHECKING, Type, Sequence
 
-from domain.move import Move
-from domain.piece import Piece
-from domain.position import Position
-from domain.side import Side
+from app.sides import White, Black
+from interface.piece import Piece
 
-location_regex = re.compile(r'^(?P<file>[a-zA-Z]+)(?P<rank>[0-9]+)$')
+if TYPE_CHECKING:
+    from interface.move import Move
+    from interface.position import Position
+    from interface.board import Board
 
 
 class King(Piece):
@@ -24,6 +23,9 @@ class King(Piece):
     def name(self):
         return "King"
 
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
+
 
 class Queen(Piece):
     @property
@@ -37,6 +39,9 @@ class Queen(Piece):
     @property
     def name(self) -> str:
         return "Queen"
+
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
 
 
 class Rook(Piece):
@@ -52,6 +57,9 @@ class Rook(Piece):
     def name(self) -> str:
         return "Rook"
 
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
+
 
 class Bishop(Piece):
     @property
@@ -65,6 +73,9 @@ class Bishop(Piece):
     @property
     def name(self) -> str:
         return 'Bishop'
+
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
 
 
 class Knight(Piece):
@@ -80,6 +91,9 @@ class Knight(Piece):
     def name(self) -> str:
         return "Knight"
 
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
+
 
 class Pawn(Piece):
     @property
@@ -94,37 +108,9 @@ class Pawn(Piece):
     def name(self) -> str:
         return "Pawn"
 
+    def available_moves(self, board: Type['Board'], position: Type['Position']) -> Sequence[Type['Move']]:
+        raise NotImplemented
 
-class WhiteSide(Side):
-    @property
-    def char(self):
-        return "w"
-
-    @property
-    def name(self) -> str:
-        return "White"
-
-    @property
-    def capitalize(self) -> bool:
-        return True
-
-
-class BlackSide(Side):
-    @property
-    def char(self):
-        return "b"
-
-    @property
-    def name(self) -> str:
-        return "Black"
-
-    @property
-    def capitalize(self) -> bool:
-        return False
-
-
-White = WhiteSide()
-Black = BlackSide()
 
 str_map = {
     'k': King,
@@ -136,127 +122,11 @@ str_map = {
 }
 
 
-def from_str(piece: str) -> Piece:
-    """
+def from_str(piece: str) -> Type['Piece']:  # TODO: Use another solution
+    """ 
     Method designed to return Piece object from given one-letter string.
     Needed only for processing FEN board-state format.
     """
     if piece.lower() not in ('k', 'q', 'r', 'b', 'n', 'p'):
         raise ValueError('"%s" not defined as Chess piece.', piece)
     return str_map[piece.lower()](White) if piece.isupper() else str_map[piece.lower()](Black)
-
-
-class StandardPosition(Position):
-    """
-    tuple with two board coordinates is too simple of course, here is a standard 2D Position object implementation.
-    """
-
-    def __init__(self, pos: Union[str, Tuple[int, int], List[int]]):  # TODO: decide to support stable and one interface
-        if isinstance(pos, tuple) or isinstance(pos, list):
-            if len(pos) != 2:
-                raise ValueError('Position should be given as tuple/list with only two ints')
-            self.__file = pos[0]
-            self.__rank = pos[1]
-        elif isinstance(pos, str):
-            output = location_regex.search(pos)
-            if not output:
-                raise ValueError('Position should be given as two letter coordinates (file, rank)')
-            self.__rank = self.__rank_from_str_to_int(output.group('rank'))
-            self.__file = self.__file_from_str_to_int(output.group('file'))
-
-    @property
-    def file(self) -> int:
-        return self.__file
-
-    @property
-    def rank(self) -> int:
-        return self.__rank
-
-    @staticmethod
-    def __rank_from_str_to_int(rank: str) -> int:
-        """
-        Converting rank from standard string format to internal int value (starts from 0 instead of 1),
-        eg. 1 in standard string means 0 in internal int format (value is a second part of position, eg "a1")
-        """
-        return int(rank) - 1
-
-    @staticmethod
-    def __rank_from_int_to_str(rank: int) -> str:
-        """
-        Converting rank from internal value to standard string format (starts from 1 instead of 0),
-        eg. 0 in internal int means 1 in standard string format (value is a second part of position, eg "a1")
-        """
-        return str(rank + 1)
-
-    @staticmethod
-    def __file_from_str_to_int(rank: str) -> int:
-        """
-        Converting file from standard string format to internal int value,
-        eg. "A" means 0, "B": 1, "Z": 25, "BA": 26 (Note: 26 == "BA", not "AA" because "A" and "AA" is an equal value,
-        just like 01 == 1 in decimal system)
-        """
-        # Warning, my own, not very well tested implementation of base26 converter
-        values = []
-        for letter in rank:
-            values.append(ascii_lowercase.index(letter.lower()))
-        index_value = 0
-        counter = 0
-        for value in reversed(values):
-            if counter < 1:
-                index_value += value
-            else:
-                index_value += (value * 26) ** counter
-            counter += 1
-        return index_value
-
-    @staticmethod
-    def __file_from_int_to_str(file: int) -> str:
-        """
-        Converting file from internal int value to standard string format,
-        eg. 0 means "A", 1: "B", 25: "Z", 26: "BA" (Note: 26 == "BA", not "AA" because "A" and "AA" is an equal value,
-        just like 01 == 1 in decimal system)
-        """
-        # Warning, my own, not very well tested implementation of base26 converter
-        output_chars = 1
-        while (len(ascii_lowercase)) ** output_chars <= file:
-            output_chars += 1
-        values = []
-        for i in range(output_chars):
-            val = (file // len(ascii_lowercase) ** i) % (len(ascii_lowercase))
-            values.append(val)
-
-        return "".join(ascii_lowercase[x] for x in reversed(values))
-
-    def __str__(self):
-        return '%s%s' % (self.__file_from_int_to_str(self.file),
-                         self.__rank_from_int_to_str(self.rank))
-
-    def __iter__(self) -> Iterator[int]:
-        for coordinate in (self.file, self.rank):
-            yield coordinate
-
-    def __getitem__(self, item) -> int:
-        if item == 0:
-            return self.file
-        elif item == 1:
-            return self.rank
-        else:
-            raise IndexError("tuple index out of range")
-
-
-class StandardMove(Move):
-    """
-    Two Position aggregator with optional pawn promotion information
-    """
-
-    @property
-    def a(self):
-        return self.__a
-
-    @property
-    def b(self):
-        return self.__b
-
-    @property
-    def promotion(self):
-        return self.__promotion
