+++ src/interface/game.py
@@ -1,10 +1,11 @@
 from abc import ABCMeta, abstractmethod
 from datetime import datetime
-from typing import Type, Tuple
+from typing import Type, Tuple, TYPE_CHECKING
 
-from domain.mode import GameMode
-from domain.player import Player
-from domain.side import Side
+if TYPE_CHECKING:
+    from interface.mode import GameMode
+    from app.player import Player
+    from interface.side import Side
 
 
 class Game(metaclass=ABCMeta):
@@ -12,7 +13,7 @@ class Game(metaclass=ABCMeta):
     Generic Game logic base class
     """
 
-    def __init__(self, player1: Player, player2: Player, mode: Type[GameMode]):
+    def __init__(self, player1: 'Player', player2: 'Player', mode: Type['GameMode']):
         self.__players = [
             player1,
             player2,
@@ -26,7 +27,7 @@ class Game(metaclass=ABCMeta):
         self.__half_moves = 0
 
     @property
-    def players(self) -> Tuple[Player, ...]:
+    def players(self) -> Tuple['Player', ...]:
         return tuple(player for _, player in self.__players.items())
 
     @property
@@ -38,7 +39,7 @@ class Game(metaclass=ABCMeta):
         return self.__create_time
 
     @property
-    def on_move(self) -> Type[Side]:
+    def on_move(self) -> Type['Side']:
         sides = tuple(self.__players.keys())
         return sides[self.__half_moves % len(self.players)]
 
@@ -46,7 +47,7 @@ class Game(metaclass=ABCMeta):
         self.__start_time = datetime.now()
 
     @abstractmethod
-    def game_state(self) -> Type[Side]:
+    def game_state(self) -> Type['Side']:
         """
         method return game state which depends on specific rules for every game mode.
         Look into "Normal" game mode class for inspirations (if is even implemented right now)
