+++ src/tests/base26_test.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from app.pieces import StandardPosition
+from app.position import StandardPosition
 
 
 class MyTestCase(TestCase):
